Automatically added on <2022-06-14T22-07:00>

+++ link.sh
@@ -25,7 +25,9 @@ function link
 		end
 	end
 	if test -f starship.toml
-		ln -s "$HOME/.dotty/starship.toml" "$HOME/.config/starship.toml"
+		if not test -f  "$HOME/.config/starship.toml"
+			ln -s "$HOME/.dotty/starship.toml" "$HOME/.config/starship.toml"
+		end
 	end
 end
